pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "selftalk"
version = "0.1.0"
description = "Self-dialogue generator CLI using Mistral Chat Completions API"
readme = "README.md"
authors = [{ name = "CTO.new AI" }]
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.25,<0.28",
  "pydantic>=2,<3",
  "typer>=0.9,<0.15",
  "rich>=13,<14",
  "python-dotenv>=1.0,<2.0",
]

[project.scripts]
selftalk = "selftalk.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=7,<9",
  "pytest-mock>=3,<4",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/selftalk"]
